# app/api/admin/employees/route.js
import { NextResponse } from "next/server";
import { audit } from "@/lib/audit";
import { prisma } from "@/lib/prisma";
import { adminGuard } from "@/lib/adminAuth";

const RESULT_LIMIT = 30;

// GET /api/admin/employees?q=...&showInactive=1 — пошук для UI скидання
// PIN / призначення ролі / детальної картки. 1900+ співробітників — тому
// без ?q повертаємо лише перші RESULT_LIMIT за іменем (просто "не пусто на
// старті"), а не весь список; пошук по externalCode/імені (contains, без
// урахування регістру) звужує. За замовчуванням деактивованих (isActive:
// false) не показуємо — showInactive=1 вмикає їх назад (для пошуку
// конкретної деактивованої людини, щоб її реактивувати).
export async function GET(request) {
  const denied = await adminGuard();
  if (denied) return denied;

  const q = request.nextUrl.searchParams.get("q")?.trim();
  const showInactive = request.nextUrl.searchParams.get("showInactive") === "1";
  const department = request.nextUrl.searchParams.get("department")?.trim();

  const employees = await prisma.employee.findMany({
    where: {
      ...(showInactive ? {} : { isActive: true }),
      ...(department ? { department } : {}),
      ...(q
        ? {
            OR: [
              { name: { contains: q, mode: "insensitive" } },
              { externalCode: { contains: q, mode: "insensitive" } },
            ],
          }
        : {}),
    },
    orderBy: { name: "asc" },
    take: RESULT_LIMIT,
    select: {
      id: true,
      name: true,
      externalCode: true,
      email: true,
      role: true,
      isActive: true,
      department: true,
      position: { select: { code: true, name: true } },
      territory: { select: { name: true } },
    },
  });

  return NextResponse.json({ employees, limit: RESULT_LIMIT });
}

// POST /api/admin/employees — ручне разове створення одного співробітника
// (окремо від масового Excel-імпорту, Фаза B — для одиничних випадків,
// коли когось треба завести негайно, не чекаючи наступного файлу). Мінімум
// обов'язкових полів: name + externalCode (той самий login-ключ, що й у
// реального імпорту, має лишатись унікальним).
export async function POST(request) {
  const denied = await adminGuard();
  if (denied) return denied;

  const body = await request.json().catch(() => ({}));
  const name = String(body.name || "").trim();
  const externalCode = String(body.externalCode || "").trim();

  if (!name) return NextResponse.json({ error: "name is required" }, { status: 400 });
  if (!externalCode) return NextResponse.json({ error: "externalCode is required" }, { status: 400 });

  const existing = await prisma.employee.findFirst({
    where: { externalCode: { equals: externalCode, mode: "insensitive" } },
  });
  if (existing) {
    return NextResponse.json({ error: "externalCode already exists" }, { status: 409 });
  }

  try {
    const created = await prisma.employee.create({
      data: {
        name,
        externalCode,
        email: body.email || null,
        department: body.department || null,
        positionId: body.positionId != null ? Number(body.positionId) : null,
        territoryId: body.territoryId != null ? Number(body.territoryId) : null,
        managerId: body.managerId != null ? Number(body.managerId) : null,
      },
      select: {
        id: true,
        name: true,
        externalCode: true,
        email: true,
        role: true,
        isActive: true,
        department: true,
        position: { select: { code: true, name: true } },
        territory: { select: { name: true } },
      },
    });
    await audit("employee.create", "employee", created.id, { externalCode: created.externalCode, name: created.name });
    return NextResponse.json(created, { status: 201 });
  } catch (err) {
    if (err.code === "P2002") {
      return NextResponse.json({ error: "email already in use" }, { status: 409 });
    }
    throw err;
  }
}