# openapi.yaml
openapi: 3.1.0
info:
  title: CarLS — Carlsberg Learning & Support API
  version: 1.0.0
  description: |
    Опис усіх Route Handlers з `app/api/**/route.*` (2026-10-04).

    **Правило:** новий або змінений ендпоінт = правка цього файлу в тому ж PR.
    `lib/openapi.test.ts` валить CI, якщо роут і цей файл розходяться
    (шлях або метод є лише з одного боку), а `oasdiff` у PR позначає
    ламаючі зміни описаних контрактів.

    Повні схеми тіл — для зовнішніх контрактів (Power Query `/api/data/*`,
    Telegram, cron, health) і для відповіді на питання / завершення модуля.
    Решта (адмінка, кабінет) — шлях, метод, авторизація й коди відповідей:
    їхні клієнти живуть у цьому ж репозиторії й перевіряються рев'ю коду.

    Перегляд — `/admin/api` (Swagger UI, лише для супер-адміна).
servers:
  - url: /
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: Сесія співробітника після входу кодом + PIN (lib/session.js). Підписана, зі sessionVersion — адмін може відкликати.
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: Окремий вхід в /admin по ADMIN_PASSWORD (lib/adminSession.js), не пов'язаний із співробітником.
    superAdminSession:
      type: apiKey
      in: cookie
      name: admin_session
      description: admin_session рівня "super" (SUPER_ADMIN_PASSWORD).
    apiToken:
      type: http
      scheme: bearer
      description: Токен Power Query, виданий в /admin (AdminApiToken). Власник має бути активним admin/hr_manager.
    cronSecret:
      type: http
      scheme: bearer
      description: CRON_SECRET — Vercel Cron підставляє його сам.
    telegramSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: TELEGRAM_WEBHOOK_SECRET, заданий при реєстрації webhook.
...
  responses:
    BadRequest:
      description: Невалідне тіло або параметри
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Forbidden:
      description: Сесія є, але роль не дозволяє
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Не знайдено (або не належить цій людині)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Конфлікт стану
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Ok:
      description: Успіх (JSON)
      content: { application/json: { schema: { type: object } } }
    Created:
      description: Створено
      content: { application/json: { schema: { type: object } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
paths:
...
  /api/admin/courses/{courseId}:
    parameters: [{ name: courseId, in: path, required: true, schema: { type: integer } }]
    get:
      tags: [admin]
      summary: Курс з модулями, екранами й компонентами для конструктора
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [admin]
      summary: Змінити налаштування курсу
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [admin]
      summary: Видалити курс
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/admin/courses/{courseId}/assign:
    post:
      tags: [admin]
      summary: Призначити курс зараз — на посади, території та/або конкретних людей
      security: [{ adminSession: [] }]
      parameters: [{ name: courseId, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                positionCodes: { type: array, items: { type: string } }
                territoryIds: { type: array, items: { type: integer } }
                employeeIds: { type: array, items: { type: integer } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { description: Помилка призначення }
...
  /api/admin/employees:
    get:
      tags: [admin]
      summary: Список співробітників (пошук, фільтри)
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Створити співробітника
      security: [{ adminSession: [] }]
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/admin/employees/{employeeId}/badges:
    parameters: [{ name: employeeId, in: path, required: true, schema: { type: integer } }]
    get:
      tags: [admin]
      summary: Відзнаки співробітника
      security: [{ adminSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Видати відзнаку співробітнику вручну
      security: [{ adminSession: [] }]
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Ця відзнака вже видана }
        "500": { description: Помилка видачі }
...
  /api/admin/badges/{badgeId}/award:
    post:
      tags: [admin]
      summary: Масова видача ручної відзнаки (посади, території, люди)
      security: [{ adminSession: [] }]
      parameters: [{ name: badgeId, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...